% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cif_curve.R
\name{check_ggsurvfit}
\alias{check_ggsurvfit}
\title{Internal: validate shapes, ranges, and emit warnings for the plot}
\usage{
check_ggsurvfit(
  survfit_object,
  lims.x,
  lims.y,
  ggsurvfit.type = NULL,
  addConfidenceInterval = TRUE,
  addCensorMark = TRUE,
  addIntercurrentEventMark = TRUE,
  shape.censor.mark = 3,
  shape.intercurrent.event.mark = 16
)
}
\description{
Internal: validate shapes, ranges, and emit warnings for the plot
}
