% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossing.R
\name{crossing}
\alias{crossing}
\title{computes at which observation a vector y crosses ref for the first time}
\usage{
crossing(yfor, ref = 0)
}
\arguments{
\item{yfor}{yfor is either a vector and a matrix}

\item{ref}{ref is the refence value}
}
\value{
whensign, a matrix with observation number at which there is crossing
}
\description{
Computes at which observation vector yfor crosses yref for the first time if it is not crossed, then 0 is returned
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
