% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffe.R
\name{diffe}
\alias{diffe}
\title{appends NA at beginning of diff(y)}
\usage{
diffe(y)
}
\arguments{
\item{y}{either a vector and a matrix}
}
\value{
Dy contains the differences of y, with NA appended at the start
}
\description{
appends NA at beginning of diff(y) and creates column names accordingly when y is either a vector and a matrix
}
\references{
Berta et al. 2020
}
\author{
P. Paruolo
}
