% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.companion.R
\name{ec.companion}
\alias{ec.companion}
\title{computes companion matrix of the VAR}
\usage{
ec.companion(est, p = 2, nlag = 4)
}
\arguments{
\item{est}{is the output of ec.EG1.R}

\item{p}{(positive integer) is the dimension of the VAR}

\item{nlag}{(positive integer) is the number of lags in the VAR}
}
\value{
mA companion matrix
}
\description{
builds the companion matrix of the VAR
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
