% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.datalag.R
\name{ec.datalag}
\alias{ec.datalag}
\title{prepares Dy y_1 Dy_1 ... Dy_{nlag-1} for estimation}
\usage{
ec.datalag(y, nlag = 4)
}
\arguments{
\item{y}{is the data matrix of variables in the VAR}

\item{nlag}{is the number of lags in the VAR (min = 2)}
}
\value{
ymat contains the folloiwing columns {Dy y_1 Dy_1 ... Dy_{nlag-1}}
}
\description{
Prepares data for estimation
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
