% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.EG1.R
\name{ec.EG1}
\alias{ec.EG1}
\title{estimates the VECM with the 2-stage procedure of Engle & Granger}
\usage{
ec.EG1(det1, det2, ymat, npl, befpn, ndet, drop1 = NA, drop2 = NA)
}
\arguments{
\item{det1}{deterministic matrix of constant(s) and trend(s)}

\item{det2}{deterministic matrix of seasonals and point dummies}

\item{ymat}{matrix of lags}

\item{npl}{n, p, nlag}

\item{befpn}{begtrim, endtrim, nforecast, npred}

\item{ndet}{order of the model d(i,j)}

\item{drop1}{selection of det1 regressors in first stage to drop}

\item{drop2}{selection of det1 regressors in second stage to drop}
}
\value{
out a list with estimates
}
\description{
Estimates the EC with EG. Cointegration rank fixed at 1
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, P.G. Lovaglio
}
