% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.gfd.R
\name{ec.gfd}
\alias{ec.gfd}
\title{plots forecasts of difference with confidence bars}
\usage{
ec.gfd(obj, whichseries = 1, nsigma = 3, xvec, yvec, cal, lar = 0.025, ...)
}
\arguments{
\item{obj}{output of ec.main}

\item{whichseries}{series number}

\item{nsigma}{how many standard deviations in confidence bars}

\item{xvec}{vector of dates to place on x axis}

\item{yvec}{vector of exp(y) values to display on y axis}

\item{cal}{calendar vector}

\item{lar}{length of arrows in error bars}

\item{...}{other plot parameters}
}
\value{
does not return output, just creates a graph
}
\description{
plots forecasts of difference with confidence bars
}
\author{
P. Paruolo,
}
