% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.gfld.R
\name{ec.gfld}
\alias{ec.gfld}
\title{ec.gfld plots forecasts of levels and difference with confidence bars}
\usage{
ec.gfld(
  obj,
  whichseries = 1,
  nsigma = 3,
  jointboo = TRUE,
  epsboo = TRUE,
  filename = "whatever",
  xvec,
  yvec,
  cal,
  lar = 0.025,
  ...
)
}
\arguments{
\item{obj}{output of ec.main}

\item{whichseries}{series number}

\item{nsigma}{how many standard deviations in confidence bars}

\item{jointboo}{boolean: TRUE if 1x2 graph, FALSE otherwise}

\item{epsboo}{boolean: TRUE eps graph, FALSE pdf graph}

\item{filename}{string, name of the file (no extension)}

\item{xvec}{vector of dates to place on x axis}

\item{yvec}{vector of exp(y) values to display on y axis}

\item{cal}{calendar vector}

\item{lar}{length of arrows in error bars}

\item{...}{other plot parameters}
}
\value{
does not return output, just creates a double graph
}
\description{
plots forecasts of levels and difference with confidence bars
}
\author{
P. Paruolo
}
