% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.ifp.R
\name{ec.ifp}
\alias{ec.ifp}
\title{Computes Indices of Forecast Performance}
\usage{
ec.ifp(afdlin, rwsigma, rwabsmean, kval = 1.959964)
}
\arguments{
\item{afdlin}{actual + forecast values + fcse}

\item{rwsigma}{standard deviation of Random Walk in sample}

\item{rwabsmean}{mean absolute deviation of Random Walk in sample}

\item{kval}{how many se to use, default kval = 1.959964}
}
\value{
list of indices of forecast performance
 1: index for model forecast
 0: index for Random Walk forecast
}
\description{
indices of forecast performance
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
