% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.plotfor.R
\name{ec.plotfor}
\alias{ec.plotfor}
\title{plots forecasts}
\usage{
ec.plotfor(
  y,
  x = NA,
  lcolact = NA,
  lcolfor = NA,
  ltypefor = NA,
  polycol = NA,
  myylim = NA,
  ...
)
}
\arguments{
\item{y}{actual values and forecasts (point forecast, lower bound, upper bound)}

\item{x}{time calendar}

\item{lcolact}{color actual (scalar)}

\item{lcolfor}{color forecasts}

\item{ltypefor}{type forecasts}

\item{polycol}{color polygons
if one wishes to have different lcolfor,ltypefor,polycol
by week > make linecol, linetype, polycol vectors, indexed by week}

\item{myylim}{vector with min and max for y axis}

\item{...}{other plot parameters}
}
\value{
does not return output, just creates a graph
}
\description{
plot actual and forecast intervals
}
\references{
Berta et al. 2020
}
\author{
P. Paruolo
}
