% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.plotroots.R
\name{ec.plotroots}
\alias{ec.plotroots}
\title{Companion matrix of the VAR}
\usage{
ec.plotroots(roots)
}
\arguments{
\item{roots}{are the roots of the companion matrix, see ec.companion.R}
}
\value{
does not return output, just creates a graph
}
\description{
plots roots and the unit circle
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
