% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec.searchbreaks.R
\name{ec.searchbreaks}
\alias{ec.searchbreaks}
\title{search for breaks dates for given q (=1,2,3,4)}
\usage{
ec.searchbreaks(qse, ymat, npl, befpn, ndet, gfillmin = 10, fixed = NA)
}
\arguments{
\item{qse}{q: number of (additional) breaks, s: start date for search, e: end date for search}

\item{ymat}{matrix of lags}

\item{npl}{n, p, nlag}

\item{befpn}{begtrim, endtrim, nforecast, npred}

\item{ndet}{order of the model d(i,j)}

\item{gfillmin}{gfill value}

\item{fixed}{vector of breaks to be taken as fixed (not between s=start and e=end)}
}
\value{
out list with break dates and values of regression average sum of squares
}
\description{
Search for location of break points in 1st-stage of Engle-Granger
}
\references{
Berta et al. 2020
}
\author{
P. Paruolo
}
