% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iculomb-data.R
\docType{data}
\name{iculomb}
\alias{iculomb}
\title{Data from Italian Civil Protection}
\format{
A dataset including 324 obs and 24 columns.
}
\source{
\url{https://github.com/pcm-dpc/COVID-19/}
}
\usage{
data(iculomb)
}
\description{
Data from Italian Civil Protection
}
\examples{
data(iculomb)

}
\references{
Italian Civil Protection
}
\keyword{datasets}
