% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagn.R
\name{lagn}
\alias{lagn}
\title{lag j of matrix or vector y}
\usage{
lagn(y, j, fill = NA)
}
\arguments{
\item{y}{column vector or matrix}

\item{j}{number of lags}

\item{fill}{value to be used to fill the missing values at the beginning, default = NA}
}
\value{
y lagged j cells, with fill in the missing j positions
}
\description{
lagn(y,j,fill=NA) produces lag j of matrix or vector y, with fill in missing j cells
             REM: alternative to "lead-lag" {dplyr} which applies to vector y
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
