% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listsize.R
\name{listsize}
\alias{listsize}
\title{listsize number of terms in the search for 1,2,3,4 number of breaks}
\usage{
listsize(myT, gfill, start)
}
\arguments{
\item{myT}{sample size}

\item{gfill}{number of gap periods}

\item{start}{beginning}
}
\value{
a vector of 4 elements, with the number of candidate models for 1,2,3,4 breaks
}
\description{
computes length-4 vector with number of terms in the search for 1,2,3,4 number of breaks
}
\author{
P. Paruolo
}
