% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cif.R
\name{summary.cif}
\alias{summary.cif}
\title{summary function for cif}
\usage{
\method{summary}{cif}(object, ..., digits = 4)
}
\arguments{
\item{object}{is the name of the cif object created by cif}

\item{...}{other parameters}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used.}
}
\value{
returns summary output from model estimation and forecasting
}
\description{
Summary function for presize
}
\references{
Berta et al. 2020
}
\author{
P. Berta, P. Paruolo, S. Verzillo, PG. Lovaglio
}
