% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wald.mls.R
\name{Wald.mls}
\alias{Wald.mls}
\title{Wald test for Multivariate Least-Squares regression}
\usage{
Wald.mls(mlsresults)
}
\arguments{
\item{mlsresults}{output of mls, mlsresults<-mls(y, x, df_flag)}
}
\value{
wald table of Wald tests on significance of single regressors and pvalues based on chi square distribution
}
\description{
Wald test for multivariate Least-Squares regression
}
\references{
Berta et al. 2020
}
\author{
P. Paruolo
}
